backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-SocketIO
Flask-Migrate
python-dotenv
eventlet
